requirement.txt
absl-py==0.12.0
aioredis==1.3.1
asgi-redis==1.2.0
asgiref==3.4.1
astunparse==1.6.3
async-timeout==4.0.1
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
cachetools==4.2.2
certifi==2020.12.5
cffi==1.15.0
channels==3.0.1
channels-redis==2.4.2
chardet==4.0.0
click==7.1.2
constantly==15.1.0
cryptography==35.0.0
daphne==3.0.2
Django==2.2.24
django-cors-headers==3.10.0
django-filter==21.1
django-rest-framework==0.1.0
django-utils-six==2.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
Flask==1.1.2
Flask-Cors==3.0.10
flatbuffers==1.12
gast==0.4.0
google-auth==1.30.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
googlemaps==4.5.3
grpcio==1.34.1
h5py==3.1.0
hiredis==2.0.0
hyperlink==21.0.0
idna==2.10
incremental==21.3.0
itsdangerous==1.1.0
Jinja2==2.11.3
kafka-python==2.0.2
Keras==2.4.3
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
Markdown==3.3.4
MarkupSafe==1.1.1
msgpack==0.6.2
msgpack-python==0.5.6
mysql-client==0.0.1
mysql-connector-python==8.0.25
mysqlclient==2.0.3
numpy==1.19.5
oauthlib==3.1.0
opencv-python==4.5.1.48
opt-einsum==3.3.0
Pillow==8.2.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==1.7.1
PyMySQL==1.0.2
pyOpenSSL==21.0.0
pytz==2021.1
PyYAML==5.4.1
redis==3.5.3
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scipy==1.6.3
service-identity==21.1.0
six==1.15.0
sqlparse==0.4.2
tensorboard==2.5.0
tensorboard-data-server==0.6.0
tensorboard-plugin-wit==1.8.0
tensorflow-estimator==2.5.0rc0
termcolor==1.1.0
Twisted==21.7.0
twisted-iocpsupport==1.0.2
txaio==21.2.1
typing-extensions==3.7.4.3
urllib3==1.26.4
Werkzeug==1.0.1
wrapt==1.12.1
zope.interface==5.4.0
